+++ mapjs/src/browser/update-connector-text.js
@@ -27,7 +27,7 @@ const createSVG = require('./create-svg'),
 			rectDOM = rectElement && rectElement[0],
 			translate = {};
 
-		let dimensions = false;
+		let textDimensions = false;
 		if (!textDOM) {
 			return false;
 		}
@@ -37,21 +37,29 @@ const createSVG = require('./create-svg'),
 		textDOM.style.fontWeight = labelTheme.text.font.weight;
 		textDOM.style.dominantBaseline = 'hanging';
 		textElement.text(labelText.trim());
-		dimensions = textDOM.getClientRects()[0];
-		translate.x = Math.round(centrePoint.x - dimensions.width / 2);
-		translate.y = Math.round(centrePoint.y - dimensions.height - 2);
-		// textDOM.style.left = Math.round(centrePoint.x - dimensions.width / 2);
-		// textDOM.style.top = Math.round(centrePoint.y - dimensions.height);
-		g[0].style.transform = `translate(${translate.x}px, ${translate.y}px)`;
-		textDOM.setAttribute('x', 0); //Math.round(centrePoint.x - dimensions.width / 2));
-		textDOM.setAttribute('y', 2); //Math.round(centrePoint.y - dimensions.height));
+		textDimensions = textDOM.getClientRects()[0];
+		let linkDimensions = parentElement[0].getElementsByClassName('mapjs-connector')[0].getClientRects()[0];
+		// Want connector label to be nearer parent node:
+		// 	Again, this is more for argmap, so may not work well in other themes.
+		let dleft = (linkDimensions.left - centrePoint.x);
+		let dright = (linkDimensions.right - centrePoint.x);
+		// translate.x number not very good, the mapjs-connector ClientRects.left and .right don't seem to relate to the actual boundary of the element.
+		// TODO: Not sure why not, might be that BoundingClientRect or BBox works better. Or might need to look up location of parent node instead.
+		// translate.x = Math.round(centrePoint.x - textDimensions.width / 2);
+		// translate.y = Math.round(centrePoint.y - textDimensions.height - 2);
+		translate.x = centrePoint.x + Math.round((Math.abs(dleft) < Math.abs(dright) ? dleft : dright)) / 100;
+		translate.y = Math.round(centrePoint.y - linkDimensions.height * 0.5);
 
-		// rectDOM.style.left = Math.round(centrePoint.x - dimensions.width / 2);
-		// rectDOM.style.top = Math.round(centrePoint.y - dimensions.height - 2);
-		rectDOM.setAttribute('x', 0); //Math.round(centrePoint.x - dimensions.width / 2));
-		rectDOM.setAttribute('y', 0); //Math.round(centrePoint.y - dimensions.height - 2));
-		rectDOM.setAttribute('height', Math.round(dimensions.height));
-		rectDOM.setAttribute('width', Math.round(dimensions.width));
+		g[0].style.transform = `translate(${translate.x}px, ${translate.y}px)`;
+		textDOM.setAttribute('x', 0); // Math.round(centrePoint.x - textDimensions.width / 2));
+		textDOM.setAttribute('y', 2); // Math.round(centrePoint.y - textDimensions.height));
+
+		// rectDOM.style.left = Math.round(centrePoint.x - textDimensions.width / 2);
+		// rectDOM.style.top = Math.round(centrePoint.y - textDimensions.height - 2);
+		rectDOM.setAttribute('x', 0); // Math.round(centrePoint.x - textDimensions.width / 2));
+		rectDOM.setAttribute('y', 0); // Math.round(centrePoint.y - textDimensions.height - 2));
+		rectDOM.setAttribute('height', Math.round(textDimensions.height));
+		rectDOM.setAttribute('width', Math.round(textDimensions.width));
 		rectDOM.style.fill = labelTheme.backgroundColor;
 		rectDOM.style.stroke = labelTheme.borderColor;
 		return textElement;